Add an image text task, datamodule, basic model
Add an image to text task and update ~innofw.utils.defaults
with default datamodule class for this task

Add the `ImageToTextDataModule` for the image to text
task. That is a PyTorch Lightning datamodule that
operates with folders that has Images folder and
captions.txt file wil comma-separated mapping between
an image and a caption.

Add the `BasicImageToTextEncoderDecoder` model which has an
encoder convolutional neural net (commonly image feature extractor)
and a decoder recurrent neural net (LSTM)

Add an experiment to train the abovementioned model on [Flickr8
dataset](https://www.kaggle.com/datasets/adityajn105/flickr8k)